requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
psutil>=5.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
scapy>=2.5.0
requests>=2.31.0
jinja2>=3.1.0